requirements.txt
Django
djangorestframework
django-cors-headers
openai
transformers
torch
python-dotenv
requests